updated imageBlock to include imageBlockView in imageBlockEdit
